requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pytest==8.0.2
geopy==2.4.1
python-multipart==0.0.9
python-dotenv==1.0.0